Add support for rendering fonts and dynamic text (#353)
This support builds on text glyph support by supporting rendering custom fonts as well as custom text through a text delegate. This enables Lottie to be used with translations as well as interactive animations.

All text operations such as fills, strokes, tracking, etc are supported by this.

Fixes #173 
Fixes #349 
Fixes #350 
#352

+++ lottie/src/main/java/com/airbnb/lottie/Font.java
@@ -7,9 +7,9 @@ class Font {
   private final String family;
   private final String name;
   private final String style;
-  private final double ascent;
+  private final float ascent;
 
-  Font(String family, String name, String style, double ascent) {
+  Font(String family, String name, String style, float ascent) {
     this.family = family;
     this.name = name;
     this.style = style;
@@ -28,7 +28,7 @@ class Font {
     return style;
   }
 
-  @SuppressWarnings("unused") double getAscent() {
+  @SuppressWarnings("unused") float getAscent() {
     return ascent;
   }
 
@@ -38,7 +38,7 @@ class Font {
       String family = json.optString("fFamily");
       String name = json.optString("fName");
       String style = json.optString("fStyle");
-      double ascent = json.optDouble("ascent");
+      float ascent = (float) json.optDouble("ascent");
       return new Font(family, name, style, ascent);
     }
   }
